이현 : 점, 원 충돌구현, 밑 작업 02. 04

+++ GameEngineBase/GameEngineTime.h
@@ -31,7 +31,7 @@ private:
 	LARGE_INTEGER startCheck_;
 	LARGE_INTEGER endCheck_;
 	double deltaTime_;
-	float Second_;
+	double Second_;
 	int F_;
 	int FPS_;
